Working header+footer+css, stub loading
