+++ public/index.html
@@ -6,6 +6,8 @@
     <meta name="theme-color" content="#000000">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico?v=201912">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons"/>
+    <link rel="stylesheet" href="https://master-7rqtwti-di4alr4iwbwyg.us-2.platformsh.site/themes/custom/neon/build/components/theme/theme.css"/>
     <title>NEON | Portal Core Components</title>
     <!-- Google Tag Manager -->
     <script>gtmDataLayer=[{'pageCategory':'Core Components'}];</script>
