+++ src/lib_components/components/NeonFooter/NeonFooter.css
@@ -1,3 +1,6 @@
+body: {
+    background-color: #ffff00 !important;
+}
 .neon-footer {
   display: grid;
   grid-gap: 2px;
