+++ src/lib_components/components/NeonPage/NeonPageHeaderFooter.css
@@ -1,1 +1,6 @@
+body: {
+    background-color: #ff0000;
+}
+/*
 @import url('https://master-7rqtwti-di4alr4iwbwyg.us-2.platformsh.site/themes/custom/neon/build/components/theme/theme.css');
+*/
